requirements.txt
cvxpy >= 1.1.15
numpy
matplotlib
pyyaml
